requirements.txt
jsonpickle==1.3
aws-xray-sdk==2.4.3
bedrock-agentcore==1.1.1
strands-agents[openai]
boto3>=1.34.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.2
python-docx>=1.1.0,<2.0.0
python-multipart>=0.0.6,<1.0.0
python-dotenv>=1.0.0
requests>=2.31.0,<3.0.0
PyJWT>=2.8.0,<3.0.0
cryptography>=41.0.0,<43.0.0
jinja2>=3.0.0
psycopg2-binary>=2.9.9
langfuse>=2.0.0
mcp>=1.0.0
httpx>=0.27.0
